requirements.txt
# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pillow>=8.0.0

# Training utilities
tqdm>=4.60.0
tensorboard>=2.5.0

# Metrics and visualization
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Optional: for better performance
opencv-python>=4.5.0
